+++ src/pipx/main.py
@@ -263,7 +263,9 @@ def run_pipx_command(args: argparse.Namespace) -> ExitCode:  # noqa: C901
             force=args.force,
         )
     elif args.command == "list":
-        return commands.list_packages(venv_container, args.include_injected, args.json, args.short)
+        return commands.list_packages(
+            venv_container, args.include_injected, args.json, args.short, args.skip_maintenance
+        )
     elif args.command == "uninstall":
         return commands.uninstall(venv_dir, constants.LOCAL_BIN_DIR, constants.LOCAL_MAN_DIR, verbose)
     elif args.command == "uninstall-all":
@@ -565,6 +567,7 @@ def _add_list(subparsers: argparse._SubParsersAction) -> None:
     g = p.add_mutually_exclusive_group()
     g.add_argument("--json", action="store_true", help="Output rich data in json format.")
     g.add_argument("--short", action="store_true", help="List packages only.")
+    g.add_argument("--skip-maintenance", action="store_true", help="Skip maintenance tasks.")
     p.add_argument("--verbose", action="store_true")